small update

+++ SRC/7zip/Compress/LZMA_Alone/ELF.h
@@ -1,6 +1,6 @@
 /* The primary ELF header. */
-typedef unsigned long int u32;
-typedef unsigned short int u16;
+typedef unsigned int u32;
+typedef unsigned short u16;
 typedef unsigned char u8;
 
 #pragma pack(push)
